feat: add overflow rules
