+++ slither/detectors/all_detectors.py
@@ -1,5 +1,5 @@
 # pylint: disable=unused-import,relative-beyond-top-level
-from .examples.backdoor import Backdoor
+from .integer_bugs.oveflow import Overflow
 from .variables.uninitialized_state_variables import UninitializedStateVarsDetection
 from .variables.uninitialized_storage_variables import UninitializedStorageVars
 from .variables.uninitialized_local_variables import UninitializedLocalVars
